backend/requirements.txt
Flask
Flask-Cors
Flask_SQLAlchemy
psycopg2-binary
web3
